move logout impl
